+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] - 2023-02-01
+
+### Changed
+
+- Allow user to specify how many addresses to keep in the "most frequent" list
+- Aggregate less frequent addresses under "others" to keep the chart readable
 ## [0.1.1] - 2023-01-26
 
 ### Changed